pkgs/molmo2_hack/requirements.txt
transformers==4.57.1
torch
torchvision
accelerate
pillow
einops
huggingface_hub
python-dotenv

